requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
asyncio==3.4.3
bitsandbytes==0.42.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
Cython==0.29.37
deprecation==2.1.0
#en-core-web-sm==3.7.1
exceptiongroup==1.2.2
filelock==3.16.1
Flask==2.0.1
fsspec==2025.2.0
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
hyperframe==6.0.1
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.0.1
joblib==1.4.2
langcodes==3.4.1
language-data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.1
nltk==3.9.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.0.3
pathlib-abc==0.1.1
pathy==0.11.0
pillow==10.4.0
pluggy==1.5.0
postgrest==0.16.11
preshed==3.0.9
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
realtime==1.0.6
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sacremoses==0.1.1
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.3
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchmocks==0.1.0
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typer==0.15.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wasabi==0.10.1
weasel==0.4.1
websockets==12.0
Werkzeug==2.0.1
zipp==3.20.2

# Include the own colorfull school-logging module from Git
# download logging package: git+https://github.com/vertok/school_logging.git@main#egg=school-logging
# install language db for spacy:python -m spacy download en_core_web_sm